requirements.txt
ttkbootstrap
deep-translator
pandas
openpyxl
pyinstaller
python-docx